Class Filter Updated
