+++ app/routes/group-finder/group-single/partials/related-groups.partial.tsx
@@ -2,7 +2,7 @@ import { Configure, InstantSearch } from "react-instantsearch";
 import { Hits } from "react-instantsearch";
 import { Button } from "~/primitives/button/button.primitive";
 import { useMemo } from "react";
-import { HitComponent } from "../../finder/components/location-search/hit-component.component";
+import { HitComponent } from "../../finder/components/hit-component.component";
 import { useLoaderData } from "react-router";
 import { LoaderReturnType } from "../loader";
 import { useResponsive } from "~/hooks/use-responsive";
